Unwrap values (except Bits) in KeyVars
Formerly most values in KeyVars were wrapped with shadow classes. Problems included:
- Strings (including enums) and numeric values were wrapper classes, making it impossible
  to look values up in dicts.
- Invalid values were of type Invalid instead of just being None; Invalid() == None
  but the usual test is "is None", and of course Invalid() fails that test.
This change fixes that by adding a "native" attribute to any class that is wrapped and can
usefully be unwrapped; this evaluates to the associated native Python data. KeyVar.set
uses this to unwrap values.

The one data type that is not unwrapped is Bits, because that already has named attributes
(one per bit field). However, Bits was changed to prohibit "native" as the name of a bit field.
No existing keyword dictionary uses that name for a bit field.

I udpated the unit test for types, but did not add a test for KeyVars.